Moved ApiResponse to different folder. Implemented ApiResponseBuilder to build success and error response
