+++ src/main/java/live/olszewski/bamboo/apiResponse/ApiResponseBuilder.java
@@ -1,0 +1,9 @@
+package live.olszewski.bamboo.apiResponse;
+
+public interface ApiResponseBuilder {
+
+    <T> ApiResponseDto<T> buildSuccessResponse(int statusCode, String message, T data);
+
+    ApiResponseDto<Void> buildErrorResponse(int statusCode, String message);
+
+}
