+++ src/main/java/live/olszewski/bamboo/apiResponse/ApiStatus.java
@@ -1,3 +1,5 @@
+package live.olszewski.bamboo.apiResponse;
+
 public enum ApiStatus {
     SUCCESS,
     ERROR
